requirements.txt
fastapi
uvicorn
boto3
slack_sdk
slack_bolt
python-dotenv
mcp
requests
urllib3